NP-285 Some further refactoring
